requirements.txt
pygame
attrs
numpy